+++ app/src/main/java/com/thenewboston/kotlinsdk/network/models/ValidatorBankDataModel.kt
@@ -8,7 +8,7 @@ data class ValidatorBankDataModel(
     @SerializedName("confirmation_expiration")
     val confirmationExpiration: Int?,
     @SerializedName("default_transaction_fee")
-    val defaultTransactionFee: Double, // 1.0
+    val defaultTransactionFee: Int, // 1
     @SerializedName("ip_address")
     val ipAddress: String, // 143.110.137.54
     @SerializedName("node_identifier")
